src/backend/requirements.txt
fastapi
sqlalchemy
pydantic
python-dotenv
alembic
uvicorn
asyncpg
psycopg2-binary
passlib
python-jose